Adapt implementer tools to the new getImplementerToolsConfig implementation. Refactor concept search box.
