+++ packages/esm-implementer-tools/src/configuration/concept-search.styles.css
@@ -1,0 +1,31 @@
+.autocomplete,
+.autocomplete > [role="combobox"] {
+  display: flex;
+  position: relative;
+  width: 20em;
+}
+
+.autocomplete [role="listbox"] {
+  margin: 0;
+  padding: 0;
+  display: flex;
+  flex-flow: column wrap;
+  background-color: whitesmoke;
+  color: #fff;
+}
+
+.autocomplete li {
+  line-height: 0.75rem;
+  padding: 0.5em;
+  display: block;
+  border-bottom: 0.125rem solid #718096;
+  outline: 0;
+  margin: 0;
+  color: #000;
+}
+
+.autocomplete [role="option"]:hover {
+  background-color: #38a169;
+  border-color: #38a169;
+  color: #f7fafc;
+}
