+++ packages/esm-implementer-tools/src/configuration/value-editor.tsx
@@ -1,18 +1,20 @@
 import React, { useEffect, useState, useRef, useCallback } from "react";
+import { ConfigValueDescriptor } from "./editable-value.component";
+import { ConceptSearchBox } from "./concept-search";
 
 interface ValueEditorProps {
-  valueString: string;
+  element: ConfigValueDescriptor;
   handleSave: (val: string) => void;
   handleClose: () => void;
 }
 
 export default function ValueEditor({
-  valueString,
+  element,
   handleSave,
   handleClose,
 }: ValueEditorProps) {
   const ref = useRef(null);
-  const [tmpValue, setTmpValue] = useState(valueString);
+  const [tmpValue, setTmpValue] = useState(JSON.stringify(element._value));
 
   const escapeListener = useCallback((e: KeyboardEvent) => {
     if (e.key === "Escape") {
@@ -40,13 +42,25 @@ export default function ValueEditor({
   }, []);
 
   return (
-    <div ref={ref}>
-      <input
-        type="text"
-        value={tmpValue}
-        onChange={(e) => setTmpValue(e.target.value)}
-      ></input>
+    <div ref={ref} style={{ display: "inherit" }}>
+      {isConcept(element) ? (
+        <ConceptSearchBox
+          setConcept={(concept) => {
+            handleSave(JSON.stringify(concept.uuid));
+          }}
+        />
+      ) : (
+        <input
+          type="text"
+          value={tmpValue}
+          onChange={(e) => setTmpValue(e.target.value)}
+        ></input>
+      )}
       <button onClick={() => handleSave(tmpValue)}>Save</button>
     </div>
   );
+}
+
+function isConcept(element) {
+  return element._value.length == 36;
 }
